confirmModal is now reusable, udpated Student Profile with this new resuable component
